Added more asmdef and references
Created Nifti and opendicom assembly references in their folders. Removed Assembly definition for the entire package and created one for Runtime and Editor separately. The Editor assembly is referencing runtime assembly. The runtime is referencing Nifti and opendicom. This set up made the app run in the editor and also allowed building an exe.